expose conv2d the same way as all other ops
